+++ ui/src/Frameworks.js
@@ -5,6 +5,7 @@ import FrameworksTable from './FrameworksTable.js';
 export default function Data(props: DataProps) {
   const [loading, setLoading] = useState(false);  
   const [frameworks, setFrameworks] = useState([]);
+  const [completed, setCompleted] = useState([]);
 
   // Function to get Apache Mesos Tasks
   const getMesosFrameworks = async () => {
@@ -13,6 +14,7 @@ export default function Data(props: DataProps) {
     const response = await fetch("http://localhost:5050/frameworks?order=dsc&limit=-1");
     const data = await response.json();
     setFrameworks(data.frameworks);
+    setCompleted(data.completed_frameworks);
     console.log(data);
     setLoading(false);
   };  
@@ -26,7 +28,12 @@ export default function Data(props: DataProps) {
       <Box>
         <div className="frameworks">
         {loading ? (<h4>Loading...</h4>) :
-          <FrameworksTable frameworks={frameworks}/>
+          <div>
+            <p></p>
+            <FrameworksTable frameworks={frameworks} title="Frameworks" />
+            <p></p>
+            <FrameworksTable frameworks={completed} title="Completed Frameworks" />
+          </div>
         }    
         </div>    
       </Box>