+++ changelog.md
@@ -1,6 +1,8 @@
 # Master
 - Open port to the mesos agent
 - Add simple webui functionality
+- Update mini Apache Mesos to version 1.11.0-0.2.0
+- Add show details of frameworks and tasks
 
 # v0.1.0
 - First inn